fix(result): Address medium priority code review issues
- Add string validation for deserialized fields to prevent memory exhaustion
  - ID and title fields limited to 256 characters
  - Description field limited to 1MB
  - Tags limited to 100 entries, each max 128 characters
- Make Results inner vector private to improve encapsulation
- Add comprehensive documentation with examples for all public APIs
- Optimize test code by avoiding unnecessary string cloning

These changes improve security, API design, and code quality without
breaking existing functionality.

+++ src/consumer/backpressure.rs
@@ -119,7 +119,7 @@ impl BackpressureController {
                     // Otherwise continue trying
                 } else {
                     // Memory limit reached, wait with exponential backoff
-                    let wait_time = Duration::from_millis(10 * (memory_wait_attempts + 1).min(100));
+                    let wait_time = Duration::from_millis((10 * (memory_wait_attempts + 1).min(100)) as u64);
                     tokio::time::sleep(wait_time).await;
                     memory_wait_attempts += 1;
                 }
